# services/wallet-api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Giano Wallet API",
    "description": "WebAuthn ceremonies, sessions and policied ERC-4337 user-operation relay for Giano smart wallets (multi-tenant).",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque session token from a ceremony verify endpoint"
      },
      "adminKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-tenant admin API key (provisioned via TENANTS_SEED)"
      }
    },
...
  },
  "paths": {
...
    "/v1/admin/sponsorship": {
      "get": {
        "summary": "Read this tenant's sponsorship rules",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "adminKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "configured": {
                      "type": "boolean"
                    },
                    "config": {},
                    "updatedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "valid": {
                      "type": "boolean"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "configured",
                    "updatedAt",
                    "valid"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Replace this tenant's sponsorship rules",
        "tags": [
          "admin"
        ],
        "description": "Full replace, validated on write. Rejected with per-path messages on any violation, so an invalid rule set is never stored and therefore never interpreted permissively.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "security": [
          {
            "adminKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "config": {},
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message",
                    "issues"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update fields of this tenant's sponsorship rules",
        "tags": [
          "admin"
        ],
        "description": "Shallow field-level merge over the stored configuration, then validated as a whole — so a partial update cannot leave the rule set in a state a full replace would have rejected.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {}
              }
            }
          }
        },
        "security": [
          {
            "adminKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "config": {},
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message",
                    "issues"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/sponsorship/history": {
      "get": {
        "summary": "Who changed the sponsorship rules, and when",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "security": [
          {
            "adminKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revisions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "config": {},
                          "createdAt": {
                            "type": "string"
                          },
                          "createdByKeyHash": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "createdAt",
                          "createdByKeyHash"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "revisions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/sponsorship/balance": {
      "get": {
        "summary": "This tenant's gas balance, reservations and where to fund",
        "tags": [
          "admin"
        ],
        "description": "Read live from the paymaster contract, with the block it was read at, so the figures can be reproduced from the chain rather than taken on trust.",
        "security": [
          {
            "adminKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "paymasterAddress": {
                      "type": "string"
                    },
                    "chainId": {
                      "type": "number"
                    },
                    "registered": {
                      "type": "boolean"
                    },
                    "enabled": {
                      "type": "boolean"
                    },
                    "withdrawAddress": {
                      "type": "string",
                      "nullable": true
                    },
                    "balanceWei": {
                      "type": "string"
                    },
                    "reservedWei": {
                      "type": "string"
                    },
                    "availableWei": {
                      "type": "string"
                    },
                    "deficitWei": {
                      "type": "string"
                    },
                    "feeWei": {
                      "type": "string"
                    },
                    "blockNumber": {
                      "type": "string"
                    },
                    "fundingInstructions": {
                      "type": "object",
                      "properties": {
                        "to": {
                          "type": "string"
                        },
                        "call": {
                          "type": "string"
                        },
                        "tenantId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "to",
                        "call",
                        "tenantId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "paymasterAddress",
                    "chainId",
                    "registered",
                    "enabled",
                    "withdrawAddress",
                    "balanceWei",
                    "reservedWei",
                    "availableWei",
                    "deficitWei",
                    "feeWei",
                    "blockNumber",
                    "fundingInstructions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}